Estimate bandwith moved to data_to_data.py
